@@ -16,9 +16,9 @@ module Middleware
 
       yield config if block_given?
 
-      raise UndefinedAuthenticatorError if config.fetch(:authenticator, nil).nil?
+      fail UndefinedAuthenticatorError if config.fetch(:authenticator, nil).nil?
       @failed_response = Rack::Response.new(
-        config[:failed_auth_message], 401, { "Content-Type" => "text/plain"}
+        config[:failed_auth_message], 401, "Content-Type" => "text/plain"
       )
     end
 
