@@ -7,9 +7,11 @@ class Client
     private
 
     def config
-      @config ||= YAML.load(File.read(File.join(
-                    Rails.root, "config", "clients.yml"))
-                  ).with_indifferent_access
+      @config ||= YAML.load(File.read(config_path)).with_indifferent_access
+    end
+
+    def config_path
+      File.join(Rails.root, "config", "clients.yml")
     end
   end
 end
