@@ -6,7 +6,7 @@ class Patient < Base
   class << self
     def search(query)
       return [] unless query
-      # TODO this should probably delegate to something like Sphinx
+      # TODO: this should probably delegate to something like Sphinx
       active.where("lower(name) like ?", "%#{query.to_s.downcase}%")
     end
   end
@@ -14,6 +14,6 @@ class Patient < Base
   private
 
   def delete_associations!
-    Case.where(patient_id: self.id).map(&:delete!)
+    Case.where(patient_id: id).map(&:delete!)
   end
 end
