@@ -1,24 +1,29 @@
-class V1::AppointmentPresenter < V1::BasePresenter
+module V1
+  class AppointmentPresenter < BasePresenter
+    SIMPLE_ATTRIBUTES = %i(
+      id
+      start
+      order
+      end
+      location
+    )
 
-  def initialize(attributes)
-    @attributes = attributes
-    @patient_attributes = attributes[:patient] || {}
-  end
-
-  def present
-    {}.tap do |presented|
-      presented[:id] = attributes["id"]
-      presented[:start] = attributes["start"]
-      presented[:order] = attributes["order"]
-      presented[:end] = attributes["end"]
-      presented[:location] = attributes["location"]
-      presented[:patient] = V1::PatientPresenter.new(patient_attributes).present
+    def initialize(attributes)
+      @attributes = attributes
+      @patient_attributes = attributes[:patient] || {}
     end
+
+    def present
+      {}.tap do |presented|
+        SIMPLE_ATTRIBUTES.each do |k|
+          presented[k] = attributes[k.to_s]
+        end
+        presented[:patient] = PatientPresenter.new(patient_attributes).present
+      end
+    end
+
+    private
+
+    attr_reader :attributes, :patient_attributes
   end
-
-
-  private
-
-  attr_reader :attributes, :patient_attributes
-
 end
