@@ -1,34 +1,33 @@
-class V1::PatientPresenter < V1::BasePresenter
+module V1
+  class PatientPresenter < BasePresenter
+    SIMPLE_ATTRIBUTES = %i(id name gender birth death)
 
-  def initialize(patient_attributes)
-    @patient_attributes = HashWithIndifferentAccess.new(patient_attributes)
-  end
+    def initialize(attributes)
+      @attributes = HashWithIndifferentAccess.new(attributes)
+    end
 
-  def present
-    {}.tap do |presented|
-      presented[:id] = patient_attributes[:id]
-      presented[:name] = patient_attributes[:name]
-      presented[:gender] = patient_attributes[:gender]
-      presented[:birth] = patient_attributes[:birth]
-      presented[:death] = patient_attributes[:death]
-
-      if patient_attributes[:cases]
-        presented[:cases] =
-          patient_attributes[:cases].map do |c|
-            {
-              id: c[:id],
-              anatomy: c[:anatomy],
-              side: c[:side]
-            }
+    # rubocop:disable Metrics/AbcSize, Metrics/MethodLength
+    def present
+      {}.tap do |presented|
+        SIMPLE_ATTRIBUTES.each do |k|
+          presented[k] = attributes[k.to_s]
+        end
+        if attributes[:cases]
+          presented[:cases] =
+            attributes[:cases].map do |c|
+              {
+                id: c[:id],
+                anatomy: c[:anatomy],
+                side: c[:side]
+              }
+            end
         end
       end
-
     end
+    # rubocop:enable Metrics/AbcSize, Metrics/MethodLength
+
+    private
+
+    attr_reader :attributes
   end
-
-
-  private
-
-  attr_reader :patient_attributes
-
 end
