@@ -10,10 +10,13 @@
 # Make sure your secret_key_base is kept private
 # if you're sharing your code publicly.
 
-# Although this is not needed for an api-only application, rails4 
-# requires secret_key_base or secret_token to be defined, otherwise an 
+# Although this is not needed for an api-only application, rails4
+# requires secret_key_base or secret_token to be defined, otherwise an
 # error is raised.
 # Using secret_token for rails3 compatibility. Change to secret_key_base
 # to avoid deprecation warning.
 # Can be safely removed in a rails3 api-only application.
+
+# rubocop:disable Metrics/LineLength
 Healer::Application.config.secret_token = '84d51184a04073c3d67278ececb64817a75f28a30dbb75c1b151ca6949e6f0cf6e59e40dbeb21c396007a38e04874f2e734147bf1245be21db6200f1798f2ab2'
+# rubocop:enable Metrics/LineLength
