@@ -14,9 +14,7 @@ RSpec.configure do |config|
 
   config.warnings = false
 
-  if config.files_to_run.one?
-    config.default_formatter = 'doc'
-  end
+  config.default_formatter = "doc" if config.files_to_run.one?
 
   config.profile_examples = 0
 
