@@ -1,4 +1,10 @@
 PATIENT_ATTRIBUTES = %i(name birth death gender)
 CASE_ATTRIBUTES = %i(anatomy side)
 APPOINTMENT_ATTRIBUTES = %i(start order location end)
-ATTACHMENT_ATTRIBUTES = %i(description document_file_name document_content_type document_file_size created_at)
+ATTACHMENT_ATTRIBUTES = %i(
+  description
+  document_file_name
+  document_content_type
+  document_file_size
+  created_at
+)
