@@ -32,11 +32,11 @@ module HTTP
     end
 
     def response_ids_for(response_records)
-      response_records.map{ |r| r["id"] }
+      response_records.map { |r| r["id"] }
     end
 
     def pluck_response_record(response_records, lookup_id)
-      response_records.detect{ |r| r["id"] == lookup_id }
+      response_records.detect { |r| r["id"] == lookup_id }
     end
   end
 
@@ -52,7 +52,7 @@ module HTTP
     def json_content_headers
       {
         "Content-Type" => "application/json",
-        "CONTENT_TYPE" => "application/json",
+        "CONTENT_TYPE" => "application/json"
       }
     end
   end
